Refactor search handling and modularize data operations
Moved `handle_search_code` logic from views to `NearbyPlaceRepository` for better encapsulation and reuse. Removed redundant constants from `NearbyPlace` and aligned their usage with the repository. This improves code organization, adheres to the repository pattern, and reduces duplication.
